requirements.txt
instagrapi==2.0.0
typer==0.7.0
rich==13.7.0
tenacity==8.2.3
python-dotenv==1.0.0
requests==2.31.0
Pillow>=8.1.1
